feat: follow-up on claude
