+++ package.json
@@ -106,7 +106,7 @@
     "braces": "in the resolutions - needed to resolve a security dependency issue - need to remove this when the issue is resolved and packages are updated"
   },
   "dependencies": {
-    "@artsy/cohesion": "4.241.0",
+    "@artsy/cohesion": "4.242.1",
     "@artsy/palette-mobile": "14.0.30",
     "@artsy/to-title-case": "1.1.0",
     "@braze/react-native-sdk": "11.0.0",
