+++ src/app/Scenes/BottomTabs/BottomTabsIcon.tsx
@@ -76,7 +76,7 @@ export const BottomTabsIcon: React.FC<{ tab: BottomTabType; state: "active" | "i
         <Path
           fillRule="evenodd"
           clipRule="evenodd"
-          d="M27.9844 39.3996C27.7027 39.3996 27.4211 39.3253 27.1863 39.1768C24.6441 37.5641 13.9878 30.5582 13.9878 22.6359C13.9878 18.4078 17.3784 15 21.5858 15C23.9193 15 26.0648 16.1384 27.5 17.9736C28.9352 16.1384 31.0807 15 33.4142 15C37.6216 15 41.0122 18.4078 41.0122 22.6359C41.0122 30.5582 30.3559 37.5641 27.8137 39.1768C27.5789 39.3253 27.2973 39.3996 27.0156 39.3996H27.9844Z"
+          d="M32.3374 16.3373C31.4033 16.3365 30.4782 16.5199 29.6149 16.8768C28.7517 17.2337 27.9672 17.7571 27.3063 18.4173L27.0041 18.7195L26.7018 18.4173C25.3675 17.0829 23.5578 16.3333 21.6707 16.3333C19.7837 16.3333 17.974 17.0829 16.6396 18.4173C15.3053 19.7516 14.5557 21.5614 14.5557 23.4484C14.5557 25.3354 15.3053 27.1452 16.6396 28.4795L26.3818 38.204C26.4616 38.289 26.558 38.3568 26.6651 38.4031C26.7721 38.4494 26.8874 38.4733 27.0041 38.4733C27.1207 38.4733 27.2361 38.4494 27.3431 38.4031C27.4501 38.3568 27.5465 38.289 27.6263 38.204L37.3685 28.4795C38.3641 27.485 39.0423 26.2174 39.3172 24.8372C39.592 23.4571 39.4512 22.0264 38.9125 20.7263C38.3738 19.4263 37.4615 18.3152 36.291 17.5339C35.1206 16.7526 33.7447 16.3362 32.3374 16.3373Z"
           fill={color("black100")}
         />
       ),
@@ -84,7 +84,7 @@ export const BottomTabsIcon: React.FC<{ tab: BottomTabType; state: "active" | "i
         <Path
           fillRule="evenodd"
           clipRule="evenodd"
-          d="M28.0312 39.3996C27.7496 39.3996 27.468 39.3253 27.2332 39.1768C24.691 37.5641 14.0347 30.5582 14.0347 22.6359C14.0347 18.4078 17.4253 15 21.6327 15C23.9662 15 26.1117 16.1384 27.5469 17.9736C28.9821 16.1384 31.1276 15 33.4611 15C37.6685 15 41.0591 18.4078 41.0591 22.6359C41.0591 30.5582 30.4028 37.5641 27.8606 39.1768C27.6258 39.3253 27.3442 39.3996 27.0625 39.3996H28.0312ZM21.6327 16.5664C18.2875 16.5664 15.5931 19.2743 15.5931 22.6359C15.5931 29.0463 23.9325 35.3067 27.531 37.5641C27.5387 37.5641 27.5551 37.5641 27.5628 37.5641C31.1614 35.3067 39.5007 29.0387 39.5007 22.6359C39.5007 19.2743 36.8064 16.5664 33.4611 16.5664C31.3183 16.5664 29.3635 17.7651 28.3398 19.6532C28.1488 19.9991 27.7972 20.2106 27.5469 20.2106C27.2882 20.2106 26.945 19.9991 26.754 19.6532C25.7304 17.7651 23.7756 16.5664 21.6327 16.5664Z"
+          d="M32.333 16.3334C31.3989 16.3326 30.4738 16.5159 29.6105 16.8728C28.7473 17.2297 27.9628 17.7532 27.3019 18.4134L26.9997 18.7156L26.6974 18.4134C25.3631 17.079 23.5534 16.3294 21.6663 16.3294C19.7793 16.3294 17.9696 17.079 16.6352 18.4134C15.3009 19.7477 14.5513 21.5574 14.5513 23.4445C14.5513 25.3315 15.3009 27.1413 16.6352 28.4756L26.3774 38.2C26.4572 38.2851 26.5536 38.3528 26.6607 38.3992C26.7677 38.4455 26.8831 38.4694 26.9997 38.4694C27.1163 38.4694 27.2317 38.4455 27.3387 38.3992C27.4457 38.3528 27.5421 38.2851 27.6219 38.2L37.3641 28.4756C38.3597 27.4811 39.0379 26.2135 39.3128 24.8333C39.5876 23.4532 39.4468 22.0225 38.9081 20.7224C38.3694 19.4223 37.4571 18.3113 36.2866 17.53C35.1162 16.7487 33.7403 16.3323 32.333 16.3334ZM36.1019 27.2134L26.9997 36.3156L17.8974 27.2134C17.3735 26.7251 16.9532 26.1363 16.6617 25.4821C16.3702 24.8279 16.2134 24.1216 16.2008 23.4055C16.1882 22.6894 16.3199 21.9781 16.5881 21.314C16.8564 20.6499 17.2556 20.0466 17.7621 19.5402C18.2685 19.0338 18.8718 18.6345 19.5359 18.3663C20.2 18.098 20.9113 17.9663 21.6274 17.9789C22.3435 17.9916 23.0497 18.1483 23.7039 18.4398C24.3582 18.7313 24.947 19.1516 25.4352 19.6756L26.3774 20.6L26.9997 21.24L27.6219 20.6L28.5641 19.6756C29.5751 18.7335 30.9124 18.2206 32.2941 18.245C33.6757 18.2694 34.994 18.8291 35.9712 19.8063C36.9484 20.7834 37.5081 22.1017 37.5325 23.4834C37.5568 24.8651 37.044 26.2023 36.1019 27.2134Z"
         />
       ),
     },
