+++ src/app/Scenes/LegacyFavorites/FavoriteArtists.tsx
