+++ src/app/Scenes/LegacyFavorites/FavoriteCategories.tsx
